mv contents of medianDetectorTest.py into diagnosticsTest.py
